Logout, age and passwords match Change match validation for passwords Add min age validation to 18 Add logout action to profile Change matches list to table Add age to profile view

+++ frontend/src/actions/registration-page-action.ts
@@ -11,6 +11,10 @@ const registrationAction = async ({ request }) => {
     const form = await request.formData();
     const registrationData = Object.fromEntries(form.entries()) as RegistrationData;
 
+    if (registrationData.password !== registrationData.confirmPassword) {
+        throw Error('Passwords do not match!');
+    }
+
     const token = await AuthService.register(registrationData);
     localStorage.setItem('token', token);
